+++ pydidas/workflow/generic_tree.py
@@ -1,6 +1,6 @@
 # This file is part of pydidas.
 #
-# Copyright 2023, Helmholtz-Zentrum Hereon
+# Copyright 2023 - 2024, Helmholtz-Zentrum Hereon
 # SPDX-License-Identifier: GPL-3.0-only
 #
 # pydidas is free software: you can redistribute it and/or modify
@@ -21,7 +21,7 @@ tree-like structure.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2023, Helmholtz-Zentrum Hereon"
+__copyright__ = "Copyright 2023 - 2024, Helmholtz-Zentrum Hereon"
 __license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Production"
@@ -417,8 +417,7 @@ class GenericTree:
         pydidas.workflow.GenericTree
             A new instance of the GenericTree
         """
-        cls = self.__class__
-        _copy = cls.__new__(cls)
+        _copy = type(self.__class__.__name__, (self.__class__,), {})()
         _copy.__dict__.update(
             {
                 _key: copy.deepcopy(_value)
@@ -429,6 +428,9 @@ class GenericTree:
                 )
             }
         )
+        if self.root is None:
+            _copy.nodes = {}
+            _copy.root = None
         if self.root is not None:
             _copy.set_root(copy.deepcopy(self.root))
         return _copy